requirements.txt
datagnosis
matplotlib==3.7.2
numpy==1.23.0
openml==0.14.1
pandas==1.4.3
PyYAML==6.0.1
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.13.2
torch==2.0.1
tqdm==4.64.0
wandb==0.15.10
xgboost==1.7.6
ipykernel
tqdm
autograd-lib 
notebook 


